Fixed Github Pages issues
Me and my friends were having hard time figuring out how we can successfully deploy our website to github. Eventhough it took us a while to get to it, it was after some time beyond the deadline. 

Please be considerate of us since we have spent a lot of time and energy to build these websites.

Thanks,